Import https://cpan.metacpan.org/authors/id/M/ML/MLEHMANN/Convert-UUlib-1.3.tar.gz

+++ uulib/crc32.c
@@ -82,7 +82,7 @@ static const crc32_t crc_table[256] = {
 };
 
 /* ========================================================================= */
-#define DO1(buf) crc = crc_table[((int)crc ^ (*buf++)) & 0xff] ^ (crc >> 8);
+#define DO1(buf) crc = crc_table[(crc ^ (*buf++)) & 0xff] ^ (crc >> 8);
 #define DO2(buf)  DO1(buf); DO1(buf);
 #define DO4(buf)  DO2(buf); DO2(buf);
 #define DO8(buf)  DO4(buf); DO4(buf);
